iupacMatrix

This function creates a iupacMatrix object to be used with the distSTRING function. By default,the iupac matrix is defined as literal distance obtained from Chang et al. 2017. (see https://link.springer.com/article/10.1007/s00335-017-9704-9)

iupacMatrix()

Value

matrix

References

Chang,P. L.,Kopania,E.,Keeble,S.,Sarver,B. A.,Larson, E.,Orth,A.,... & Dean,M. D. (2017). Whole exome sequencing of wild-derived inbred strains of mice improves power to link phenotype and genotype. Mammalian genome,28(9-10),416-425.

See also

Author

Kristian K Ullrich

Examples

iupacMatrix()

#>      A    C    G    T    R    Y    S    W    K    M  B  D  H  V  .  -  N  X
#> A  0.0  1.0  1.0  1.0  0.5  1.0  1.0  0.5  1.0  0.5 -1 -1 -1 -1 -1 -1 -1 -1
#> C  1.0  0.0  1.0  1.0  1.0  0.5  0.5  1.0  1.0  0.5 -1 -1 -1 -1 -1 -1 -1 -1
#> G  1.0  1.0  0.0  1.0  0.5  1.0  0.5  1.0  0.5  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> T  1.0  1.0  1.0  0.0  1.0  0.5  1.0  0.5  0.5  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> R  0.5  1.0  0.5  1.0  0.0  1.0  1.0  1.0  1.0  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> Y  1.0  0.5  1.0  0.5  1.0  0.0  1.0  1.0  1.0  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> S  1.0  0.5  0.5  1.0  1.0  1.0  0.0  1.0  1.0  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> W  0.5  1.0  1.0  0.5  1.0  1.0  1.0  0.0  1.0  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> K  1.0  1.0  0.5  0.5  1.0  1.0  1.0  1.0  0.0  1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> M  0.5  0.5  1.0  1.0  1.0  1.0  1.0  1.0  1.0  0.0 -1 -1 -1 -1 -1 -1 -1 -1
#> B -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> D -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> H -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> V -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> . -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> - -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> N -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1
#> X -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1.0 -1 -1 -1 -1 -1 -1 -1 -1